+++ src/BetterSafe.cpp
@@ -14,67 +14,54 @@ void BetterSafe::loadSafe(
     std::function<void()> success,
     std::function<void(int)> failure
 ) {
-    if (!getSafeLevels(type).empty()) return success();
+    if (!safes[type].empty()) return success();
 
     listener->bind([failure = std::move(failure), success = std::move(success), type](web::WebTask::Event* e) {
         if (auto res = e->getValue()) {
             if (!res->ok()) return failure(res->code());
 
-            auto json = res->json().unwrapOr(matjson::Value::array());
-            if (!json.isArray()) return success();
-
-            getSafeLevels(type) = ranges::map<std::vector<SafeLevel>>(json.asArray().unwrap(), [type](const matjson::Value& level) {
-                return SafeLevel {
-                    .id = (int)level["id"].asInt().unwrapOr(0),
-                    .levelID = (int)level["levelID"].asInt().unwrapOr(0),
-                    .dates = level.contains("dates") && level["dates"].isArray() ? ranges::map<std::vector<SafeDate>>(
-                        level["dates"].asArray().unwrap(),
-                        [](const matjson::Value& date) {
-                            auto parts = string::split(date.asString().unwrapOr("1970-01-01"), "-");
-                            return SafeDate {
-                                .year = (uint16_t)(parts.size() > 0 ? numFromString<int>(parts[0]).unwrapOr(1970) : 1970),
-                                .month = (uint8_t)(parts.size() > 1 ? numFromString<int>(parts[1]).unwrapOr(1) : 1),
-                                .day = (uint8_t)(parts.size() > 2 ? numFromString<int>(parts[2]).unwrapOr(1) : 1)
-                            };
-                        }
-                    ) : std::vector<SafeDate>(),
-                    .type = type,
-                    .tier = (int)level["tier"].asInt().unwrapOr(0)
-                };
-            });
+            safes[type] = ranges::map<std::vector<SafeLevel>>(
+                res->json().unwrapOrDefault().asArray().unwrapOrDefault(),
+                [type](const matjson::Value& v) {
+                    SafeLevel level;
+                    GEODE_UNWRAP_INTO_IF_OK(level.id, v.get("id").andThen([](const matjson::Value& v) { return v.as<int>(); }));
+                    GEODE_UNWRAP_INTO_IF_OK(level.levelID, v.get("levelID").andThen([](const matjson::Value& v) { return v.as<int>(); }));
+                    GEODE_UNWRAP_INTO_IF_OK(level.dates, v.get("dates").andThen([](const matjson::Value& v) {
+                        return v.asArray().map([](const std::vector<matjson::Value>& vec) {
+                            return ranges::map<std::vector<SafeDate>>(vec, [](const matjson::Value& d) {
+                                SafeDate date;
+                                if (!d.isString()) return date;
+                                auto parts = string::split(d.asString().unwrap(), "-");
+                                if (parts.size() < 1) return date;
+                                GEODE_UNWRAP_INTO_IF_OK(date.year, numFromString<int>(parts[0]));
+                                if (parts.size() < 2) return date;
+                                GEODE_UNWRAP_INTO_IF_OK(date.month, numFromString<int>(parts[1]));
+                                if (parts.size() < 3) return date;
+                                GEODE_UNWRAP_INTO_IF_OK(date.day, numFromString<int>(parts[2]));
+                                return date;
+                            });
+                        });
+                    }));
+                    level.type = type;
+                    GEODE_UNWRAP_INTO_IF_OK(level.tier, v.get("tier").andThen([](const matjson::Value& v) { return v.as<int>(); }));
+                    return level;
+                }
+            );
 
             success();
         }
     });
 
-    auto url = "";
-    switch (type) {
-        case GJTimedLevelType::Daily: url = DAILY_SAFE_URL; break;
-        case GJTimedLevelType::Weekly: url = WEEKLY_SAFE_URL; break;
-        case GJTimedLevelType::Event: url = EVENT_SAFE_URL; break;
-        default: return;
-    }
-    listener->setFilter(web::WebRequest().get(url));
+    constexpr std::array urls = { DAILY_SAFE_URL, WEEKLY_SAFE_URL, EVENT_SAFE_URL };
+    listener->setFilter(web::WebRequest().get(urls[(int)type]));
 }
 
 std::vector<SafeLevel> BetterSafe::getMonth(int year, int month, GJTimedLevelType type) {
     using SafeLevels = std::vector<SafeLevel>;
 
-    return ranges::reduce<SafeLevels>(getSafeLevels(type), [year, month](SafeLevels& levels, const SafeLevel& level) {
+    return ranges::reduce<SafeLevels>(safes[type], [year, month](SafeLevels& levels, const SafeLevel& level) {
         ranges::push(levels, ranges::reduce<SafeLevels>(level.dates, [&level, year, month](SafeLevels& dates, const SafeDate& date) {
             if (date.year == year && date.month == month) dates.push_back(level);
         }));
     });
-}
-
-std::vector<SafeLevel>& BetterSafe::getSafeLevels(GJTimedLevelType type) {
-    switch (type) {
-        case GJTimedLevelType::Daily: return DAILY_SAFE;
-        case GJTimedLevelType::Weekly: return WEEKLY_SAFE;
-        case GJTimedLevelType::Event: return EVENT_SAFE;
-        default: {
-            static std::vector<SafeLevel> empty;
-            return empty;
-        }
-    }
 }