+++ src/classes/BSCalendarPopup.cpp
@@ -129,7 +129,7 @@ bool BSCalendarPopup::setup(CCObject* obj, SEL_MenuHandler onSafe, GJTimedLevelT
         m_firstButton->setVisible(false);
         m_lastButton->setVisible(false);
         m_monthButton->setEnabled(false);
-        BetterSafe::getSafeLevels(m_type).clear();
+        BetterSafe::safes[m_type].clear();
         loadSafe(true);
     });
     refreshButton->setPosition({ 340.0f, 80.0f });
@@ -147,7 +147,9 @@ void BSCalendarPopup::closeHoverNode() {
     if (m_hoverNode) {
         auto [year, month, day] = m_currentDay;
         m_hoverNode->close();
-        m_currentDay = { year, month, day };
+        m_currentDay.year = year;
+        m_currentDay.month = month;
+        m_currentDay.day = day;
     }
 }
 
@@ -161,7 +163,7 @@ void BSCalendarPopup::createWeekdayLabel(const char* text, int idx) {
 
 void BSCalendarPopup::loadSafe(bool refresh) {
     BetterSafe::loadSafe(m_type, &m_listener, [this, refresh] {
-        auto& safe = BetterSafe::getSafeLevels(m_type);
+        auto& safe = BetterSafe::safes[m_type];
         if (safe.empty()) return;
 
         auto timeinfo = fmt::localtime(time(0));
@@ -186,12 +188,14 @@ void BSCalendarPopup::loadSafe(bool refresh) {
 }
 
 void BSCalendarPopup::loadMonth(int year, int month, bool refresh) {
+    if (month < 1 || month > 12) return;
+
     m_year = year;
     m_month = month;
 
     closeHoverNode();
     m_calendarMenu->removeAllChildren();
-    m_monthLabel->setString(fmt::format("{} {}", BSSelectPopup::MONTHS[m_month - 1], m_year).c_str());
+    m_monthLabel->setString(fmt::format("{} {}", BSSelectPopup::months[m_month - 1], m_year).c_str());
     m_monthButton->updateSprite();
     m_monthButton->setEnabled(false);
     m_prevButton->setVisible(false);
@@ -203,7 +207,7 @@ void BSCalendarPopup::loadMonth(int year, int month, bool refresh) {
     if (levelSafe.empty()) return setupMonth(nullptr);
 
     auto searchObject = GJSearchObject::create(SearchType::MapPackOnClick, ranges::reduce<std::string>(levelSafe,
-        [](std::string& str, const SafeLevel& level) { str += (str.empty() ? "" : ",") + std::to_string(level.levelID); }));
+        [](std::string& str, const SafeLevel& level) { str += (str.empty() ? "" : ",") + fmt::to_string(level.levelID); }));
 
     auto glm = GameLevelManager::get();
     if (!refresh) {
@@ -243,36 +247,39 @@ void BSCalendarPopup::setupMonth(CCArray* levels) {
 
     tm timeinfo = { 0, 0, 0, 1, m_month - 1, m_year - 1900 };
     auto firstWeekday = fmt::localtime(mktime(&timeinfo)).tm_wday;
-    if (!Mod::get()->getSettingValue<bool>("sunday-first")) firstWeekday = (firstWeekday + 6) % 7;
+    auto mod = Mod::get();
+    if (!mod->getSettingValue<bool>("sunday-first")) firstWeekday = (firstWeekday + 6) % 7;
 
     auto levelSafe = BetterSafe::getMonth(m_year, m_month, m_type);
-    auto daysInMonth = 0;
-    switch (m_month) {
-        case 1: case 3: case 5: case 7: case 8: case 10: case 12: daysInMonth = 31; break;
-        case 2: daysInMonth = 28 + (m_year % 4 == 0 && (m_year % 100 != 0 || m_year % 400 == 0)); break;
-        case 4: case 6: case 9: case 11: daysInMonth = 30; break;
-    }
+    auto levelMap = ranges::reduce<std::unordered_map<int, GJGameLevel*>>(CCArrayExt<GJGameLevel*>(levels),
+        [](std::unordered_map<int, GJGameLevel*>& acc, GJGameLevel* obj) {
+            acc[obj->m_levelID.value()] = obj;
+        });
+
+    constexpr std::array daysPerMonth = { 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31 };
+    auto daysInMonth = m_month <= daysPerMonth.size() ? daysPerMonth[m_month - 1] : 0;
+    if (m_month == 2 && (m_year % 4 == 0 && (m_year % 100 != 0 || m_year % 400 == 0))) daysInMonth++;
+
+    auto showCheckmarks = mod->getSettingValue<bool>("show-checkmarks");
+    auto spriteFrameCache = CCSpriteFrameCache::get();
     for (int i = 0; i < daysInMonth; i++) {
-        auto safeLevelIt = std::ranges::find_if(levelSafe, [this, i](const SafeLevel& level) {
+        auto safeLevelIt = std::ranges::find_if(levelSafe, [this, i = i + 1](const SafeLevel& level) {
             return std::ranges::any_of(level.dates, [this, i](const SafeDate& date) {
-                return date.year == m_year && date.month == m_month && date.day == i + 1;
+                return date.year == m_year && date.month == m_month && date.day == i;
             });
         });
         if (safeLevelIt == levelSafe.end()) continue;
 
         auto& safeLevel = *safeLevelIt;
-        auto gameLevels = reinterpret_cast<GJGameLevel**>(levels->data->arr);
-        auto gameLevelsEnd = gameLevels + levels->count();
-        auto gameLevelIt = std::find_if(gameLevels, gameLevelsEnd, [&safeLevel](GJGameLevel* level) {
-            return level->m_levelID.value() == safeLevel.levelID;
-        });
-        if (gameLevelIt == gameLevelsEnd) continue;
+        auto gameLevelIt = levelMap.find(safeLevel.levelID);
+        if (gameLevelIt == levelMap.end()) continue;
 
-        auto gameLevel = *gameLevelIt;
-
+        auto gameLevel = gameLevelIt->second;
         auto levelDifficulty = gameLevel->m_demon > 0 ? gameLevel->m_demonDifficulty > 0 ? gameLevel->m_demonDifficulty + 4 : 6 :
             gameLevel->m_autoLevel ? -1 : gameLevel->m_ratings < 5 ? 0 : gameLevel->m_ratingsSum / gameLevel->m_ratings;
-        auto diffFrame = levelDifficulty == -1 ? "diffIcon_auto_btn_001.png" : fmt::format("diffIcon_{:02d}_btn_001.png", levelDifficulty);
+        auto diffFrame = levelDifficulty == -1 ? "diffIcon_auto_btn_001.png" : fmt::format("diffIcon_{:02}_btn_001.png", levelDifficulty);
+        auto featureState = gameLevel->m_featured > 0 ? gameLevel->m_isEpic + 1 : 0;
+
         if (auto moreDifficulties = Loader::get()->getLoadedMod("uproxide.more_difficulties")) {
             auto legacy = moreDifficulties->getSettingValue<bool>("legacy-difficulties");
             auto levelStars = gameLevel->m_stars.value();
@@ -284,71 +291,56 @@ void BSCalendarPopup::setupMonth(CCArray* levels) {
                 diffFrame = fmt::format("uproxide.more_difficulties/MD_Difficulty09Small{}.png", legacy ? "_Legacy" : "");
         }
 
-        auto hasBetweenDemon = false;
+        CCPoint featurePosition;
         if (auto demonsInBetween = Loader::get()->getLoadedMod("hiimjustin000.demons_in_between")) {
             if (demonsInBetween->getSettingValue<bool>("enable-difficulties") && safeLevel.tier > 0) {
-                auto betweenDifficulty = 0;
-                switch (safeLevel.tier) {
-                    case 1: betweenDifficulty = 1; break;
-                    case 2: betweenDifficulty = 2; break;
-                    case 3: betweenDifficulty = 3; break;
-                    case 4: betweenDifficulty = 4; break;
-                    case 5: betweenDifficulty = 5; break;
-                    case 6: betweenDifficulty = 6; break;
-                    case 7: betweenDifficulty = 7; break;
-                    case 8: betweenDifficulty = 8; break;
-                    case 9: betweenDifficulty = 9; break;
-                    case 10: betweenDifficulty = 10; break;
-                    case 11: case 12: betweenDifficulty = 11; break;
-                    case 13: betweenDifficulty = 12; break;
-                    case 14: betweenDifficulty = 13; break;
-                    case 15: case 16: betweenDifficulty = 14; break;
-                    case 17: case 18: betweenDifficulty = 15; break;
-                    case 19: betweenDifficulty = 16; break;
-                    case 20: betweenDifficulty = 17; break;
-                    case 21: betweenDifficulty = 18; break;
-                    case 22: betweenDifficulty = 19; break;
-                    case 23: case 24: case 25: case 26: case 27: case 28: case 29:
-                    case 30: case 31: case 32: case 33: case 34: case 35: betweenDifficulty = 20; break;
-                }
-                if (betweenDifficulty > 0) {
-                    diffFrame = fmt::format("hiimjustin000.demons_in_between/DIB_{:02d}_001.png", betweenDifficulty);
-                    hasBetweenDemon = true;
+                constexpr std::array difficulties = {
+                    0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 11, 12, 13, 14, 14, 15, 15,
+                    16, 17, 18, 19, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20
+                };
+                if (safeLevel.tier < difficulties.size()) {
+                    diffFrame = fmt::format("hiimjustin000.demons_in_between/DIB_{:02}_001.png", difficulties[safeLevel.tier]);
+                    featurePosition.y = 9.5f;
                 }
             }
         }
 
-        auto completedLevel = GameStatsManager::get()->hasCompletedLevel(gameLevel);
+        auto completedLevel = showCheckmarks && GameStatsManager::get()->hasCompletedLevel(gameLevel);
 
         auto diffIcon = CCSprite::createWithSpriteFrameName(diffFrame.c_str());
         diffIcon->setScale(0.75f);
         diffIcon->setOpacity(255 - (completedLevel * 105));
 
-        if (gameLevel->m_featured > 0) {
-            auto featureFrame = "";
-            switch (gameLevel->m_isEpic) {
-                case 0: featureFrame = "GJ_featuredCoin_001.png"; break;
-                case 1: featureFrame = "GJ_epicCoin_001.png"; break;
-                case 2: featureFrame = "GJ_epicCoin2_001.png"; break;
-                case 3: featureFrame = "GJ_epicCoin3_001.png"; break;
+        if (featurePosition.y > 0.0f) featurePosition.x = diffIcon->getContentWidth() / 2.0f;
+        else featurePosition = diffIcon->getContentSize() / 2.0f - CCPoint { 0.0f, 5.5f };
+
+        if (featureState > 0) {
+            CCSprite* featureIcon = nullptr;
+            if (Loader::get()->isModLoaded("uproxide.animated_fire") && featureState > 1) {
+                constexpr std::array prefixes = {
+                    "uproxide.animated_fire/EpicFrame_",
+                    "uproxide.animated_fire/LegendaryFrame_",
+                    "uproxide.animated_fire/MythicFrame_"
+                };
+                auto prefix = prefixes[featureState - 2];
+                auto spriteFrames = CCArray::create();
+                for (int i = 1; i < 9; i++) {
+                    spriteFrames->addObject(spriteFrameCache->spriteFrameByName(fmt::format("{}{:02}.png", prefix, i).c_str()));
+                }
+                featureIcon = CCSprite::createWithSpriteFrame(static_cast<CCSpriteFrame*>(spriteFrames->objectAtIndex(0)));
+                featureIcon->runAction(CCRepeatForever::create(CCAnimate::create(CCAnimation::createWithSpriteFrames(spriteFrames, 1.0f / 12.0f))));
+                featurePosition.y += 15.875f + (featureState == 2);
             }
-            if (auto featureIcon = CCSprite::createWithSpriteFrameName(featureFrame)) {
-                featureIcon->setPosition(diffIcon->getContentSize() / 2.0f + CCPoint { 0.0f, -5.5f });
-                if (hasBetweenDemon) featureIcon->setPositionY(9.5f);
+            else {
+                constexpr std::array frames = {
+                    "GJ_featuredCoin_001.png", "GJ_epicCoin_001.png", "GJ_epicCoin2_001.png", "GJ_epicCoin3_001.png"
+                };
+                featureIcon = CCSprite::createWithSpriteFrameName(featureState <= frames.size() ? frames[featureState - 1] : "");
+            }
+            if (featureIcon) {
+                featureIcon->setPosition(featurePosition);
                 featureIcon->setOpacity(255 - (completedLevel * 105));
                 diffIcon->addChild(featureIcon, -2);
-                if (gameLevel->m_isEpic == 3 && !completedLevel) {
-                    auto mythicParticles = GameToolbox::particleFromString(
-                        "30a-1a1.3a0.2a20a90a0a10a5a20a20a0a0a8a0a0a0a4a1a0a0a0a0a1a0a1a0a1a0a1a1a0a0a0a0a0.784314"
-                        "a0a1a0a1a0a0.27a0a0.27a0a0a0a0a0a0a0a0a2a1a0a0a0a0a0a0a0.25a0a0a0a0a0a0a0a0a0a0a0",
-                        nullptr, false
-                    );
-                    mythicParticles->setPosition(featureIcon->getPosition());
-                    mythicParticles->setScale(0.9f);
-                    mythicParticles->resetSystem();
-                    diffIcon->addChild(mythicParticles, -1);
-                    for (int _ = 0; _ < 5; _++) mythicParticles->update(0.15f);
-                }
             }
         }
 
@@ -356,6 +348,13 @@ void BSCalendarPopup::setupMonth(CCArray* levels) {
             auto completedIcon = CCSprite::createWithSpriteFrameName("GJ_completesIcon_001.png");
             completedIcon->setPosition(diffIcon->getContentSize() / 2.0f);
             diffIcon->addChild(completedIcon, 1);
+        }
+        else {
+            auto dayLabel = CCLabelBMFont::create(fmt::to_string(i + 1).c_str(), "bigFont.fnt");
+            dayLabel->setScale(0.4f);
+            dayLabel->setPosition({ diffIcon->getContentWidth() + 3.0f, -3.0f });
+            dayLabel->setAnchorPoint({ 1.0f, 0.0f });
+            diffIcon->addChild(dayLabel, 1);
         }
 
         auto hoverButton = CCMenuItemExt::createSpriteExtra(diffIcon, [this, i, gameLevel, safeLevel](CCMenuItemSpriteExtra* sender) {
@@ -365,7 +364,9 @@ void BSCalendarPopup::setupMonth(CCArray* levels) {
                 if (year == m_year && month == m_month && i + 1 == day) return;
             }
             createHoverNode(sender, safeLevel, gameLevel);
-            m_currentDay = { (uint16_t)m_year, (uint8_t)m_month, (uint8_t)(i + 1) };
+            m_currentDay.year = m_year;
+            m_currentDay.month = m_month;
+            m_currentDay.day = i + 1;
         });
         hoverButton->setPosition({ (i + firstWeekday) % 7 * 38.0f + 36.0f, 219.0f - (int)((i + firstWeekday) / 7) * 38.0f });
         hoverButton->setID(fmt::format("level-button-{}", i + 1));
@@ -373,13 +374,100 @@ void BSCalendarPopup::setupMonth(CCArray* levels) {
 
         if (m_year == m_currentDay.year && m_month == m_currentDay.month && i + 1 == m_currentDay.day)
             createHoverNode(hoverButton, safeLevel, gameLevel);
+
+        if (completedLevel || featureState < 0 || featureState > 4) continue;
+
+        auto RP = Loader::get()->getLoadedMod("sawblade.rating_particles");
+        if (RP) {
+            constexpr std::array toggles = {
+                "show-rate", "show-featured", "show-epic", "show-legendary", "show-mythic"
+            };
+            if (!RP->getSettingValue<bool>(toggles[featureState])) continue;
+        }
+        else if (featureState != 4) continue;
+
+        constexpr std::array colors = {
+            "rate-color", "featured-color", "epic-color", "legendary-color", "mythic-color"
+        };
+        auto color = colors[featureState];
+
+        auto particles = CCParticleSystemQuad::createWithTotalParticles(RP ? RP->getSettingValue<int64_t>("amount") : 30, false);
+        particles->setDuration(-1.0f);
+        particles->setLife(1.3f);
+        particles->setLifeVar(0.2f);
+        particles->setEmissionRate(RP ? RP->getSettingValue<int64_t>("amount") : 20.0f);
+        particles->setAngle(90.0f);
+        particles->setAngleVar(0.0f);
+        particles->setSpeed(RP ? RP->getSettingValue<int64_t>("speed") : 10.0f);
+        particles->setSpeedVar(5.0f);
+        particles->setPosVar({ 20.0f, 20.0f });
+        particles->setGravity({ 0.0f, 0.0f });
+        particles->setRadialAccel(8.0f);
+        particles->setRadialAccelVar(0.0f);
+        particles->setTangentialAccel(0.0f);
+        particles->setTangentialAccelVar(0.0f);
+        particles->setStartSize(RP ? RP->getSettingValue<int64_t>("size") : 20.0f);
+        particles->setStartSizeVar(1.0f);
+        particles->setStartSpin(0.0f);
+        particles->setStartSpinVar(0.0f);
+        particles->setStartColor(to4F(to4B(RP ? RP->getSettingValue<ccColor3B>(color) : ccColor3B { 0, 255, 255 })));
+        particles->setStartColorVar({ 0.0f, 0.0f, 0.0f, 0.0f });
+        particles->setEndSize(1.0f);
+        particles->setEndSizeVar(1.0f);
+        particles->setEndSpin(0.0f);
+        particles->setEndSpinVar(0.0f);
+        particles->setEndColor(to4F(to4B(RP ? lighten3B(RP->getSettingValue<ccColor3B>(color), 30) : ccColor3B { 0, 200, 255 })));
+        particles->setEndColorVar({ 0.0f, 0.0f, 0.0f, 0.0f });
+        particles->m_fFadeInTime = 0.27f;
+        particles->m_fFadeInTimeVar = 0.0f;
+        particles->m_fFadeOutTime = 0.27f;
+        particles->m_fFadeOutTimeVar = 0.0f;
+        particles->setStartRadius(0.0f);
+        particles->setStartRadiusVar(0.0f);
+        particles->setEndRadius(0.0f);
+        particles->setEndRadiusVar(0.0f);
+        particles->setRotatePerSecond(0.0f);
+        particles->setRotatePerSecondVar(0.0f);
+        particles->setEmitterMode(kCCParticleModeGravity);
+        particles->setPositionType(kCCPositionTypeGrouped);
+        particles->setBlendAdditive(RP ? RP->getSettingValue<bool>("blending") : true);
+        particles->m_bStartSpinEqualToEnd = false;
+        particles->setRotationIsDir(false);
+        particles->m_bDynamicRotationIsDir = false;
+        particles->m_uParticleIdx = RP ? RP->getSettingValue<int64_t>("texture") : 0;
+        particles->setDisplayFrame(spriteFrameCache->spriteFrameByName(fmt::format("particle_{:02}_001.png", particles->m_uParticleIdx).c_str()));
+        particles->m_bUseUniformColorMode = false;
+        particles->m_fFrictionPos = 0.0f;
+        particles->m_fFrictionPosVar = 0.25f;
+        particles->m_fRespawn = 0.0f;
+        particles->m_fRespawnVar = 0.0f;
+        particles->m_bOrderSensitive = false;
+        particles->m_bStartSizeEqualToEnd = false;
+        particles->m_bStartRadiusEqualToEnd = false;
+        particles->m_bStartRGBVarSync = false;
+        particles->m_bEndRGBVarSync = false;
+        particles->m_fFrictionSize = 0.0f;
+        particles->m_fFrictionSizeVar = 0.0f;
+        particles->m_fFrictionRot = 0.0f;
+        particles->m_fFrictionRotVar = 0.0f;
+        particles->setPosition(featurePosition);
+        particles->setScale(0.9f);
+        particles->resetSystem();
+        diffIcon->addChild(particles, -1);
+        particles->update(0.15f);
+        particles->update(0.15f);
+        particles->update(0.15f);
+        particles->update(0.15f);
+        particles->update(0.15f);
     }
 }
 
 void BSCalendarPopup::createHoverNode(CCMenuItemSpriteExtra* sender, const SafeLevel& safeLevel, GJGameLevel* gameLevel) {
     m_hoverNode = BSHoverNode::create(safeLevel, gameLevel, [this] {
         m_hoverNode = nullptr;
-        m_currentDay = {};
+        m_currentDay.year = 0;
+        m_currentDay.month = 0;
+        m_currentDay.day = 0;
     });
     m_hoverNode->setPosition(sender->getPosition() + CCPoint {
         0.0f,